Remove tree-sitter-vue
tree-sitter-vue [1] is not compatible with Node.js 22. This is because
it uses tree-sitter [2] v0.19.x, which does not support Node.js 22.

The version of tree-sitter that works with Node.js 22 is v0.21.0 or
later, as it uses the new Node-API [3] instead of NAN [4] for its Node
bindings.

Since tree-sitter-vue hasn't been updated to use v0.21.0+, it doesn't
work with Node.js 22.

There is no official grammar for Vue that uses tree-sitter v0.21+ and
Node-API yet.

[1] https://github.com/ikatyang/tree-sitter-vue
[2] https://github.com/tree-sitter/tree-sitter
[3] https://nodejs.github.io/node-addon-examples/about/what/
[4] https://github.com/nodejs/nan